Fixed some bugs.
When dirty extents list is empty, some special treatments take place.
And the interval value of a extent from extent tree can not be changed.

Added some dump functions when MBUG.

Please note that there is a race bug when running on a server. But it
can not be reproduced on a server whose hardware is not so good.